Extract render shaders to another file
